+++ packages/sources/cryptocompare-test/src/endpoint/index.ts
@@ -1,0 +1,1 @@
+export { endpoint as cryptoEndpoint } from './crypto'