+++ packages/composites/historical-average/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/historical-average-adapter
 
+## 1.2.25
+
+### Patch Changes
+
+- Updated dependencies [367e6c179]
+  - @chainlink/coinmarketcap-adapter@1.5.25
+
 ## 1.2.24
 
 ### Patch Changes